Move `grpcpp/support/status_code_enum.h` to `grpcpp/impl/status.h`
