+++ BUILD
@@ -317,7 +317,6 @@ GRPCXX_PUBLIC_HDRS = [
     "include/grpc++/support/config.h",
     "include/grpc++/support/slice.h",
     "include/grpc++/support/status.h",
-    "include/grpc++/support/status_code_enum.h",
     "include/grpc++/support/string_ref.h",
     "include/grpc++/support/stub_options.h",
     "include/grpc++/support/sync_stream.h",
@@ -391,7 +390,6 @@ GRPCXX_PUBLIC_HDRS = [
     "include/grpcpp/support/server_interceptor.h",
     "include/grpcpp/support/slice.h",
     "include/grpcpp/support/status.h",
-    "include/grpcpp/support/status_code_enum.h",
     "include/grpcpp/support/string_ref.h",
     "include/grpcpp/support/stub_options.h",
     "include/grpcpp/support/sync_stream.h",
@@ -420,7 +418,6 @@ GRPCXX_PUBLIC_HDRS = [
     "include/grpc++/impl/codegen/server_interface.h",
     "include/grpc++/impl/codegen/service_type.h",
     "include/grpc++/impl/codegen/slice.h",
-    "include/grpc++/impl/codegen/status_code_enum.h",
     "include/grpc++/impl/codegen/status.h",
     "include/grpc++/impl/codegen/string_ref.h",
     "include/grpc++/impl/codegen/stub_options.h",
@@ -464,7 +461,6 @@ GRPCXX_PUBLIC_HDRS = [
     "include/grpcpp/impl/codegen/server_interface.h",
     "include/grpcpp/impl/codegen/service_type.h",
     "include/grpcpp/impl/codegen/slice.h",
-    "include/grpcpp/impl/codegen/status_code_enum.h",
     "include/grpcpp/impl/codegen/status.h",
     "include/grpcpp/impl/codegen/string_ref.h",
     "include/grpcpp/impl/codegen/stub_options.h",
