--- include/grpcpp/impl/codegen/status_code_enum.h
@@ -1,27 +1,0 @@
-/*
- *
- * Copyright 2016 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-#ifndef GRPCPP_IMPL_CODEGEN_STATUS_CODE_ENUM_H
-#define GRPCPP_IMPL_CODEGEN_STATUS_CODE_ENUM_H
-
-// IWYU pragma: private, include <grpcpp/support/status.h>
-
-/// TODO(chengyuc): Remove this file after solving compatibility.
-#include <grpcpp/support/status_code_enum.h>
-
-#endif  // GRPCPP_IMPL_CODEGEN_STATUS_CODE_ENUM_H
